refactor(widgets): fold F0Widget into the experimental Widget
There were two widget cards drawing the same thing: F0Widget (title only, with
its own drag handle and overflow menu) and the experimental Widget (headers with
subtitles, counts, links, summaries). Rather than keep both and have them
disagree about what a widget header holds, Widget absorbs what only F0Widget
could do:

- draggable + onDragStart/onDragEnd/isDragging, with the same Handle glyph and
  data-gs-handle attribute a gridstack board looks for
- selected, and an actions overflow menu
- AIButton

F0Widget is now an alias of Widget so the export keeps resolving; its stories go
(Widget has its own) and DashboardWidget renders Widget directly.

Home drops the drag chrome it had to hand-roll: SortableWidget reports the
sortable state through a render callback and SlotWidget forwards it to Widget, so
the handle is the design systems own.
