+++ packages/react/src/sds/Home/WidgetContainer/SortableWidget.tsx
@@ -3,24 +3,28 @@ import { type CSSProperties, ReactNode } from "react"
 import { useSortable } from "@dnd-kit/sortable"
 import { CSS } from "@dnd-kit/utilities"
 
-import { F0Icon } from "@/components/F0Icon"
-import { Sort } from "@/icons/app"
-import { cn } from "@/lib/utils"
+/** What the sortable state hands to the widget being rendered. */
+export interface SortableWidgetState {
+  /** Show the widget's own drag handle. False for a locked widget. */
+  draggable: boolean
+  isDragging: boolean
+}
 
 export interface SortableWidgetProps {
-  /** Classes for the drag handle, so it matches the widget header controls. */
-  handleClassName?: string
   id: string
   /** A locked widget can neither be picked up nor displaced by another. */
   disabled?: boolean
-  children: ReactNode
+  children: (state: SortableWidgetState) => ReactNode
 }
 
 /**
- * One draggable widget in an editable column. The drag surface is a HANDLE in
- * the widget's top-left — styled like its header controls — rather than the
- * whole card: a widget body is full of links, and making all of it draggable
- * would swallow them.
+ * One draggable widget in an editable column. It contributes no drag chrome of
+ * its own: the f0 `Widget` already draws a handle beside its title when
+ * `draggable`, so this reports the sortable state through its render callback
+ * and lets the design system own the affordance.
+ *
+ * dnd-kit's listeners sit on a wrapper rather than on the handle itself — the
+ * handle lives inside `Widget`, so the events reach it by bubbling.
  *
  * A `disabled` (locked) widget is neither draggable nor a drop target, so a
  * pinned widget can't be picked up OR pushed out of its place.
@@ -28,7 +32,6 @@ export interface SortableWidgetProps {
 export const SortableWidget = ({
   id,
   disabled = false,
-  handleClassName,
   children,
 }: SortableWidgetProps) => {
   const {
@@ -43,27 +46,20 @@ export const SortableWidget = ({
   const style: CSSProperties = {
     transform: CSS.Translate.toString(transform),
     transition,
-    // The dragged card rides above its neighbours and dims slightly, so the gap
-    // it will land in stays readable underneath.
+    // The dragged card rides above its neighbours so the gap it will land in
+    // stays readable underneath.
     zIndex: isDragging ? 10 : undefined,
-    opacity: isDragging ? 0.85 : undefined,
     position: "relative",
   }
 
   return (
-    <div ref={setNodeRef} style={style}>
-      {children}
-      {disabled ? null : (
-        <button
-          type="button"
-          aria-label="Reorder widget"
-          className={cn(handleClassName, "cursor-grab active:cursor-grabbing")}
-          {...attributes}
-          {...listeners}
-        >
-          <F0Icon size="sm" icon={Sort} className="text-f1-icon-bold" />
-        </button>
-      )}
+    <div
+      ref={setNodeRef}
+      style={style}
+      {...(disabled ? {} : attributes)}
+      {...(disabled ? {} : listeners)}
+    >
+      {children({ draggable: !disabled, isDragging })}
     </div>
   )
 }
